+++ cmd/gpu_plugin/gpu_plugin_test.go
@@ -61,7 +61,7 @@ func (m *mockResourceManager) GetPreferredFractionalAllocation(*v1beta1.Preferre
 	return &v1beta1.PreferredAllocationResponse{}, &dpapi.UseDefaultMethodError{}
 }
 
-func (m *mockResourceManager) SetTileCountPerCard(counts []uint64) {
+func (m *mockResourceManager) SetTileCountPerCard(counts []int) {
 }
 
 func createTestFiles(root string, devfsdirs, sysfsdirs []string, sysfsfiles map[string][]byte) (string, string, error) {
